computer-use/shadowbox/pyproject.toml
[project]
name = "shadowbox"
version = "0.1.0"
description = "Screen recording, video analysis, and CUA automation"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
    "pynput>=1.7.6",
    "pyautogui>=0.9.54",
    "Pillow>=11.1.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.26.0",
    "azure-cognitiveservices-speech>=1.34.0",
    "azure-core>=1.30.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["shadowbox*"]
